Some of the routes are changed and improved

+++ routes/recipe.js
@@ -5,9 +5,34 @@ const User = require('../models/User');
 
 router.post('/create', async (req, res) => {
   try {
-    const newRecipe = new Recipe(req.body);
-    await newRecipe.save();
+    const { userId } = req.cookies;
+    if (!userId) {
+      return res.status(401).json({ message: 'User not authenticated' });
+    }
+
+    const user = await User.findById(userId);
+    if (!user) {
+      return res.status(400).json({ message: 'User not found' });
+    }
+
+    const newRecipe = new Recipe({
+      ...req.body,
+      createdBy: userId,
+    });
+
+    const savedRecipe = await newRecipe.save();
+    console.log('Saved recipe', savedRecipe);
+
     res.status(200).json({ message: 'Recipe created successfully' });
+  } catch (error) {
+    res.status(400).json({ error: error.message });
+  }
+});
+router.get('/user/:userId/recipes', async (req, res) => {
+  try {
+    const recipes = await Recipe.find({createdBy: req.params.userId});
+    console.log('Fetched recipes', recipes); 
+    res.status(200).json(recipes);
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
@@ -15,6 +40,15 @@ router.post('/create', async (req, res) => {
 
 router.put('/update/:id', async (req, res) => {
   try {
+    const recipe = await Recipe.findById(req.params.id);
+    const { userId } = req.cookies;
+    console.log('User ID from cookies:', userId);
+    console.log('Recipe createdBy:', recipe.createdBy.toString());
+
+    if (recipe.createdBy.toString() !== userId) {
+      return res.status(403).json({ message: 'User not authorized to edit this recipe' });
+    }
+
     const updatedRecipe = await Recipe.findByIdAndUpdate(req.params.id, req.body, { new: true });
     res.status(200).json(updatedRecipe);
   } catch (error) {
@@ -22,14 +56,24 @@ router.put('/update/:id', async (req, res) => {
   }
 });
 
+
 router.delete('/delete/:id', async (req, res) => {
   try {
+    const recipe = await Recipe.findById(req.params.id);
+    const { userId } = req.cookies;
+    console.log('Recipe deleted successfully');
+
+    if (recipe.createdBy.toString() !== userId) {
+      return res.status(403).json({ message: 'User not authorized to delete this recipe' });
+    }
+
     await Recipe.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: 'Recipe deleted successfully' });
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
 });
+
 
 
 router.get('/', async (req, res) => {
@@ -96,4 +140,6 @@ router.post('/:id/rate', async (req, res) => {
 });
 
 
+
 module.exports = router;
+
